Add book Mindset by Carol Dweck
It's a great book about how we can learn the right mindset to overcome
obstacles and take in challenges. Daniel Pink mentions Carol Dweck's
work in his great book Drive, which made me want to read it. Found that
by understanding the way our brain works and how we behave when faced
with everyday challenges, we can reason about it and become ever more
resilient. Antifragile even (which is another great book, by the way).